+++ src/type_builder.js
@@ -452,6 +452,27 @@ var TypeBuilder = Object.create(null, {
     writable: false,
     enumerable: true,
     configurable: false
+  },
+
+  /**
+   * @memberOf {class4js.TypeException}
+   * @static
+   * @public
+   * @method getArgumentNames
+   * @param {Function} func
+   * @returns {Array} 
+   */
+  getArgumentNames: {
+    value: function (func) {
+      return func.toString()
+		    .match(/^[\s\(]*function[^(]*\(([^)]*)\)/)[1]
+		    .replace(/\/\/.*?[\r\n]|\/\*(?:.|[\r\n])*?\*\//g, '')
+		    .replace(/\s+/g, '')
+		    .split(',');
+    },
+    writable: false,
+    enumerable: true,
+    configurable: false
   }
 
 });