+++ tests/browser/scripts/enum.js
@@ -3,22 +3,22 @@ $run('Enum', function () {
   'use strict';
 
   var Priority = $enum({
-    low: 0,
-    normal: 1,
-    high: 2
+    LOW: 0,
+    NORMAL: 1,
+    HIGH: 2
   });
 
   var index = 0;
   for (var field in Priority) {
     switch(index) {
       case 0:
-        $assert(field == 'low');
+        $assert(field == 'LOW');
         break;
       case 1:
-        $assert(field == 'normal');
+        $assert(field == 'NORMAL');
         break;
       case 2:
-        $assert(field == 'high');
+        $assert(field == 'HIGH');
         break;
     }
     index++;