+++ tests/node/extension.js
@@ -48,6 +48,6 @@ collection.forEach(function (item) {
 });
 
 // It's required for PhantomJS
-if (typeof phantom !== "undefined") {
+if (typeof phantom !== 'undefined') {
   phantom.exit();
 }